refactor: add MovementTablePlugin
